HWKAPM-620 Refactor client modules to separate collector/instrumentation classes from the more generic concepts that can be used in other clients (e.g. opentracing)
