+++ client/collector/src/main/java/org/hawkular/apm/client/collector/SessionManager.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.apm.client.api;
+package org.hawkular.apm.client.collector;
 
 import org.hawkular.apm.api.model.trace.Node;
