--- client/collector/src/main/resources/META-INF/services/org.hawkular.apm.client.api.TraceCollector
@@ -1,1 +1,0 @@
-org.hawkular.apm.client.collector.DefaultTraceCollector
