+++ client/collector/src/main/resources/META-INF/services/org.hawkular.apm.client.collector.TraceCollector
@@ -1,0 +1,1 @@
+org.hawkular.apm.client.collector.internal.DefaultTraceCollector
