+++ client/instrumenter/src/main/java/org/hawkular/apm/instrumenter/RuleHelper.java
@@ -27,10 +27,10 @@ import org.hawkular.apm.api.logging.Logger.Level;
 import org.hawkular.apm.api.model.config.Direction;
 import org.hawkular.apm.api.model.trace.Node;
 import org.hawkular.apm.api.services.ServiceResolver;
-import org.hawkular.apm.client.api.HeadersAccessor;
-import org.hawkular.apm.client.api.SessionManager;
-import org.hawkular.apm.client.api.TraceCollector;
+import org.hawkular.apm.client.collector.SessionManager;
+import org.hawkular.apm.client.collector.TraceCollector;
 import org.hawkular.apm.instrumenter.faults.FaultDescriptor;
+import org.hawkular.apm.instrumenter.headers.HeadersAccessor;
 import org.hawkular.apm.instrumenter.io.InstrumentedInputStream;
 import org.hawkular.apm.instrumenter.io.InstrumentedOutputStream;
 import org.jboss.byteman.rule.Rule;
