+++ client/instrumenter/src/main/java/org/hawkular/apm/instrumenter/headers/JbossSwitchyardHeadersAccessor.java
@@ -22,7 +22,6 @@ import java.util.Map;
 
 import org.hawkular.apm.api.logging.Logger;
 import org.hawkular.apm.api.logging.Logger.Level;
-import org.hawkular.apm.client.api.HeadersAccessor;
 
 /**
  * The headers accessor implementation for JBoss Switchyard.
