--- client/instrumenter/src/test/resources/META-INF/services/org.hawkular.apm.client.api.HeadersFactory
@@ -1,1 +1,0 @@
-org.hawkular.apm.instrumenter.TestHeadersFactory
