+++ tests/instrumentation-framework/src/main/resources/apm-env-variables.properties
@@ -26,4 +26,4 @@ HAWKULAR_APM_PASSWORD=password
 
 HAWKULAR_APM_LOG_LEVEL=FINEST
 
-JAVA_OPTS=-javaagent:/opt/hawkular-apm-agent/hawkular-apm-agent.jar=boot:/opt/hawkular-apm-agent/hawkular-apm-agent.jar -Djboss.modules.system.pkgs=org.jboss.byteman,org.hawkular.apm.instrumenter,org.hawkular.apm.client.api
+JAVA_OPTS=-javaagent:/opt/hawkular-apm-agent/hawkular-apm-agent.jar=boot:/opt/hawkular-apm-agent/hawkular-apm-agent.jar -Djboss.modules.system.pkgs=org.jboss.byteman,org.hawkular.apm.instrumenter,org.hawkular.apm.client.collector
